pyproject.toml
[project]
name = "agent-trust"
version = "0.1.0"
description = "MCP Server for Agent Reputation & Trust Scoring"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli]>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic-settings>=2.0.0",
    "redis[hiredis]>=5.0.0",
    "structlog>=24.0.0",
    "pynacl>=1.5.0",
    "pyjwt>=2.8.0",
    "arq>=0.25.0",
    "starlette>=0.40.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
    "httpx-sse>=0.4.0",
]

[project.scripts]
agent-trust = "agent_trust.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "hypothesis>=6.0.0",
    "factory-boy>=3.3.0",
    "ruff>=0.4.0",
    "pytest-mock>=3.14.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_trust"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
